# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Host:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      hostname:
        type: string
      id:
        type: integer
      ip:
        type: string
      os:
        type: string
      ports:
        items:
          $ref: '#/definitions/models.Port'
        type: array
      teamID:
        type: string
      updatedAt:
        type: string
    type: object
  models.LoginReq:
    properties:
      password:
        type: string
      user:
        type: string
    type: object
  models.Port:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      hostID:
        type: integer
      id:
        type: integer
      number:
        type: integer
      protocol:
        type: string
      service:
        type: string
      state:
        type: string
      updatedAt:
        type: string
    type: object
  models.RegisterReq:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
  models.Scan:
    properties:
      endTime:
        type: string
      hosts:
        items:
          $ref: '#/definitions/models.Host'
        type: array
      startTime:
        type: string
      status:
        type: string
    type: object
  models.Team:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      hosts:
        items:
          $ref: '#/definitions/models.Host'
        type: array
      id:
        type: integer
      iprange:
        type: string
      name:
        type: string
      tid:
        type: string
      updatedAt:
        type: string
    type: object
  models.UserReq:
    properties:
      active:
        type: boolean
      roles:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: Login Data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Login
      tags:
      - user
  /auth/logout:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: Logout
      tags:
      - user
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a user
      parameters:
      - description: Login Data
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models.RegisterReq'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Register User
      tags:
      - user
  /auth/status:
    get:
      consumes:
      - application/json
      description: Check login status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Auth Status
      tags:
      - user
  /auth/user/{uid}:
    delete:
      consumes:
      - application/json
      description: delete a user
      parameters:
      - description: User ID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: delete user
      tags:
      - user
  /auth/users:
    get:
      consumes:
      - application/json
      description: List users on the sytem
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      summary: List users
      tags:
      - user
  /auth/users/{uid}:
    put:
      consumes:
      - application/json
      description: update users attributes
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserReq'
      - description: User ID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: update user
      tags:
      - user
  /health:
    get:
      consumes:
      - application/json
      description: Health check
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health Check
      tags:
      - status
  /hosts/by-team/:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get all hosts by team
      tags:
      - hosts
  /hosts/by-team/{tid}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: tid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get hosts by team
      tags:
      - hosts
  /jobs:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get jobs
      tags:
      - jobs
  /jobs/{jobtype}/next:
    get:
      consumes:
      - application/json
      parameters:
      - description: Job Type
        in: path
        name: jobtype
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: delete team
      tags:
      - jobs
  /jobs/manager:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get jobmanager state
      tags:
      - jobs
  /jobs/nmap/{jid}:
    post:
      consumes:
      - application/json
      description: upload new nmap scan
      parameters:
      - description: nmap scan data
        in: body
        name: scan
        required: true
        schema:
          $ref: '#/definitions/models.Scan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: nmap scan
      tags:
      - jobs
  /teams:
    get:
      consumes:
      - application/json
      description: get all teams
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: create a team
      parameters:
      - description: team data
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/models.Team'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create team
      tags:
      - teams
  /teams/{tid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: tid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: delete team
      tags:
      - teams
swagger: "2.0"